vault-embedder/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vault-embedder"
version = "0.1.0"
description = "Local vault indexer using bge-micro-v2 embeddings with Qdrant"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Marc", email = "marc@blitzmed.com"}
]
keywords = ["obsidian", "vault", "embeddings", "semantic-search", "qdrant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "sentence-transformers>=2.2.0",
    "qdrant-client>=1.7.0",
    "pyyaml>=6.0",
    "markdown-it-py>=3.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pathspec>=0.11.0",
    "xxhash>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
vault-embedder = "src.cli:main"

[project.urls]
"Homepage" = "https://github.com/bmbsystemsdir/openclaw-unified-plugins"
"Bug Tracker" = "https://github.com/bmbsystemsdir/openclaw-unified-plugins/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
